# Conflicts:
#	TimerSkipList.h

+++ BaseTimerSkipList.h
@@ -9,8 +9,8 @@
 
 class BaseTimerSkipList : public BaseSkipList{
 public:
-    virtual double getLastInsertOrUpdateTime() const = 0;
-    virtual double getLastPeekTime() const = 0;
+    virtual long getLastInsertOrUpdateTime() const = 0;
+    virtual long getLastPeekTime() const = 0;
     virtual double getAverageInsertOrUpdateTime() const = 0;
     virtual double getAveragePeekTime() const = 0;
 };
